Renamed 'db' directory to 'management'
