+++ src/engine/ecs/management/component_batch.cpp
