+++ tests/engine/ecs/component_set_signature_test.hpp
@@ -1,7 +1,7 @@
 #ifndef __CHESTNUT_COMPONENT_SET_SIGNATURE_TEST_H__
 #define __CHESTNUT_COMPONENT_SET_SIGNATURE_TEST_H__
 
-#include "engine/ecs/db/component_set_signature.hpp"
+#include "engine/ecs/management/component_set_signature.hpp"
 #include "engine/ecs/components/components.hpp"
 
 namespace chestnut
