+++ tests/engine/ecs/component_vector_wrapper_test.hpp
@@ -1,7 +1,7 @@
 #ifndef __CHESTNUT_COMPONENT_VECTOR_WRAPPER_TEST_H__
 #define __CHESTNUT_COMPONENT_VECTOR_WRAPPER_TEST_H__
 
-#include "engine/ecs/db/component_vector_wrapper.hpp"
+#include "engine/ecs/management/component_vector_wrapper.hpp"
 #include "engine/ecs/components/components.hpp"
 
 namespace chestnut
